forward declare ClientContext in ClientContextWrapper, so ClientContextWrapper can be safely included in templated classes
