+++ src/include/duckdb/main/client_context.hpp
@@ -289,22 +289,4 @@ private:
 	lock_guard<mutex> client_guard;
 };
 
-class ClientContextWrapper {
-public:
-	explicit ClientContextWrapper(const shared_ptr<ClientContext> &context)
-	    : client_context(context) {
-
-	      };
-	shared_ptr<ClientContext> GetContext() {
-		auto actual_context = client_context.lock();
-		if (!actual_context) {
-			throw ConnectionException("Connection has already been closed");
-		}
-		return actual_context;
-	}
-
-private:
-	weak_ptr<ClientContext> client_context;
-};
-
 } // namespace duckdb
