+++ src/main/CMakeLists.txt
@@ -19,6 +19,7 @@ add_library_unity(
   attached_database.cpp
   client_context_file_opener.cpp
   client_context.cpp
+  client_context_wrapper.cpp
   client_data.cpp
   client_verify.cpp
   connection_manager.cpp
