+++ src/main/client_context_wrapper.cpp
@@ -1,0 +1,16 @@
+#include "duckdb/main/client_context_wrapper.hpp"
+#include "duckdb/main/client_context.hpp"
+
+namespace duckdb {
+
+ClientContextWrapper::ClientContextWrapper(const shared_ptr<ClientContext> &context) : client_context(context) {};
+
+shared_ptr<ClientContext> ClientContextWrapper::GetContext() {
+	auto actual_context = client_context.lock();
+	if (!actual_context) {
+		throw ConnectionException("Connection has already been closed");
+	}
+	return actual_context;
+}
+
+} // namespace duckdb
